Refactor moving routes endpoint to routes.py
